Switch context iff on user-space
I guess it works if the timer-interrupt raised on the user-space, but
the reason why I said "I guess" is printf is a IO-intensive job and the
provided example showed the user-thread is working correctly even on
this type of job. Thus, I'm going to find a way how to apply this on
kernel area.